refactor: rename base error file for clarity
* Rename the module to better reflect that it contains a
  single base error class.
* Update imports to keep the project structure consistent.

This change improves code readability and makes error modules
easier to navigate.
