+++ exceptions/base_exception.py
